fixed-login-erros
the login error was apparently set with my database, where the password was not matching the code I wrote (it being a simple string and the actual password hatched).

Changed the logic where the password can be read as hatched (**) or normal string password.

After re-starting the server and client sides, and making sure the .env file has the right credentials and the right setting.

And finally checking all the ports, CORS, and setting to make sure backend runs in 5003 and the frontend 3000