managers-page-basic-setting
basic structure and feature for the managers page.

Still with errors, but the basic structure is set, I need to start for the next branch to work on the deeper structure and connection in between the features of the page.

We have so far the 2 optional schedules for the manager to choose from.

The list of employees available per day and hour at the bottom of them page.

And the middle table calendar where they can make there on schedules and save.